+++ middlewares/celebrateValidation.js
@@ -1,0 +1,69 @@
+const { celebrate, Joi } = require('celebrate');
+const isURL = require('validator/lib/isURL');
+
+const urlValidation = (value, helpers) => {
+  if (isURL(value)) {
+    return value;
+  }
+  return helpers.message('Неверно переданы данные URL');
+};
+
+const validationGetUser = celebrate({
+  body: Joi.object().keys({
+    name: Joi.string().min(2).max(30).required(),
+    email: Joi.string().email().max(30).required(),
+  }),
+});
+
+const validationUpdateUserInfo = celebrate({
+  body: Joi.object().keys({
+    name: Joi.string().min(2).max(30).required(),
+    email: Joi.string().email().required(),
+  }),
+});
+
+const validationCreateMovie = celebrate({
+  body: Joi.object().keys({
+    country: Joi.string().required(),
+    director: Joi.string().required(),
+    duration: Joi.number().integer().required(),
+    year: Joi.string().required(),
+    description: Joi.string().required(),
+    image: Joi.string().required().custom(urlValidation),
+    trailer: Joi.string().required().custom(urlValidation),
+    thumbnail: Joi.string().required().custom(urlValidation),
+    nameRU: Joi.string().required(),
+    nameEN: Joi.string().required(),
+    movieId: Joi.number().integer().required(),
+  }),
+});
+
+const validationDeleteMovie = celebrate({
+  params: Joi.object().keys({
+    movieId: Joi.string().hex().length(24),
+  }),
+});
+
+const validationLogin = celebrate({
+  body: Joi.object().keys({
+    email: Joi.string().email().required(),
+    password: Joi.string().required().min(8),
+  }),
+});
+
+const validationCreateUser = celebrate({
+  body: Joi.object().keys({
+    name: Joi.string().min(2).max(30),
+    email: Joi.string().email().required(),
+    password: Joi.string().required().min(8),
+  }),
+});
+
+module.exports = {
+  validationGetUser,
+  validationUpdateUserInfo,
+  validationCreateMovie,
+  validationDeleteMovie,
+  validationLogin,
+  validationCreateUser,
+};
