+++ routes/auth.js
@@ -1,23 +1,9 @@
 const authRouter = require('express').Router();
-const { celebrate, Joi } = require('celebrate');
+const { login, createUser } = require('../controllers/users');
+const { validationLogin, validationCreateUser } = require('../middlewares/celebrateValidation');
 
-const {
-  login, createUser,
-} = require('../controllers/users');
+authRouter.post('/signin', validationLogin, login);
 
-authRouter.post('/signin', celebrate({
-  body: Joi.object().keys({
-    email: Joi.string().email().required(),
-    password: Joi.string().required().min(8),
-  }),
-}), login);
-
-authRouter.post('/signup', celebrate({
-  body: Joi.object().keys({
-    name: Joi.string().min(2).max(30),
-    email: Joi.string().email().required(),
-    password: Joi.string().required().min(8),
-  }),
-}), createUser);
+authRouter.post('/signup', validationCreateUser, createUser);
 
 module.exports = authRouter;
