+++ routes/movies.js
@@ -1,39 +1,14 @@
 const movieRouter = require('express').Router();
-const { celebrate, Joi } = require('celebrate');
-const isURL = require('validator/lib/isURL');
 const auth = require('../middlewares/auth');
 const {
   getMovies, createMovie, deleteMovie,
 } = require('../controllers/movies');
+const { validationCreateMovie, validationDeleteMovie } = require('../middlewares/celebrateValidation');
 
-const urlValidation = (value, helpers) => {
-  if (isURL(value)) {
-    return value;
-  }
-  return helpers.message('Неверно переданы данные URL');
-};
 movieRouter.get('/movies', auth, getMovies);
 
-movieRouter.post('/movies', auth, celebrate({
-  body: Joi.object().keys({
-    country: Joi.string().required(),
-    director: Joi.string().required(),
-    duration: Joi.number().integer().required(),
-    year: Joi.string().required(),
-    description: Joi.string().required(),
-    image: Joi.string().required().custom(urlValidation),
-    trailer: Joi.string().required().custom(urlValidation),
-    thumbnail: Joi.string().required().custom(urlValidation),
-    nameRU: Joi.string().required(),
-    nameEN: Joi.string().required(),
-    movieId: Joi.number().integer().required(),
-  }),
-}), createMovie);
+movieRouter.post('/movies', auth, validationCreateMovie, createMovie);
 
-movieRouter.delete('/movies/:movieId', auth, celebrate({
-  params: Joi.object().keys({
-    movieId: Joi.string().hex().length(24),
-  }),
-}), deleteMovie);
+movieRouter.delete('/movies/:movieId', auth, validationDeleteMovie, deleteMovie);
 
 module.exports = movieRouter;
